Реализован логика для экрана в классе GameViewModel.kt

+++ app/src/main/java/com/example/composition/domain/entity/GameSettings.kt
@@ -9,5 +9,5 @@ data class GameSettings (
     val maxSumValue: Int,
     val minCountOfRightAnswers: Int,
     val minPercentOfRightAnswers: Int,
-    val gameTimeInSeconds: Int
+    val gameTimeInSeconds: Long
 ): Parcelable
